requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
cffi==2.0.0
constantly==23.10.4
cryptography==46.0.7
daphne==4.0.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
openpyxl==3.1.2
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-decouple==3.8
pytz==2026.1.post1
service-identity==24.2.0
setuptools==82.0.1
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
wheel==0.46.3
zope.interface==8.2
